@@ -1,4 +1,6 @@
-﻿namespace Presentation.Shared.States;
+﻿using Presentation.Shared.Common.Abstractions;
+
+namespace Presentation.Shared.States;
 
 public class BurgerMenuState : IState
 {
