@@ -1,10 +1,8 @@
 ﻿@page "/faq"
 
-
 <MudContainer MaxWidth="MaxWidth.Large">
 
-  <MudText Typo="Typo.h1" Class="mb-10">Питання та відповіді</MudText>
-
+  <MudText Typo="Typo.h1" Class="mb-10">Питання та відповіді </MudText>
   <MudExpansionPanels Class="mb-24">
     <MudExpansionPanel>
       <TitleContent>
