@@ -1,8 +1,8 @@
-﻿<!-- System -->
-@using System.Net.Http
+﻿@using System.Net.Http
 @using System.Net.Http.Json
+@using System.Diagnostics
+@using MudBlazor
 
-<!-- Microsoft -->
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
@@ -10,21 +10,22 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 
-<!-- MudBlazor -->
-@using MudBlazor
-
-<!-- Presentation -->
 @using Presentation
-@using Presentation.Client
-@using Presentation.Client.Layout
 @using Presentation.Components
+@using Presentation.Components.Layout
 
-<!-- Presentation.Shared -->
-@using Presentation.Components.Shared.Widgets
 
-<!-- Shared -->
+@using Presentation.Client
+@using Presentation.Client.Components.Footer
+@using Presentation.Client.Components.Footer.Partials
+@using Presentation.Components.Header
+@using Presentation.Components.Header.Partials
+@using Presentation.Client.Components.Shared.Widgets
+
 @using Presentation.Shared.Common
-@using Presentation.Shared.Theme;
+@using Presentation.Shared.Common.States
+@using Presentation.Shared.Theme
+@using Presentation.Shared.States
+@using Presentation.Shared.Common.Routing
 
-<!-- inject -->
 @inject StateContainer State
